Test: Updates to make a basic test work
This commit adds one test that attempts to render the Filter component
and then simulate a click on a radio button, with the expectation that
the callback passed from the parent is called with the right value.

In order for .getByLabelText to work, the <label> element in filter.js
needed to have the 'for' property updated to reference the id of the
actual radio buttons.